intercept onLoad and add values before setValue
This solution workins in a manner that makes setValue work again.
Before setValue was run when the store fired the load event and
the custom values we wanted to add was added to late because of that.
By intercepting the onLoad function on the combo before we can add these
to the store.